requirements.txt
torch
torchvision
numpy
pillow
streamlit
requests